Update for Laravel 5 and move to PSR-4
