+++ composer.json
@@ -11,14 +11,14 @@
     ],
     "require": {
         "php": ">=5.4.0",
-        "illuminate/support": "~4.2",
+        "illuminate/support": "~5.0",
         "jwhennessey/phpinsight": "~2.0"
     },
     "require-dev": {
         "phpunit/phpunit": "~4.0"
     },
     "autoload": {
-        "psr-0": {
+        "psr-4": {
             "Antoineaugusti\\LaravelSentimentAnalysis\\": "src/"
         }
     }
