+++ src/Facades/SentimentAnalysis.php
