+++ src/SentimentAnalysis.php
@@ -4,18 +4,23 @@ use PHPInsight\Sentiment;
 
 class SentimentAnalysis {
 
+	/**
+	 * @var \PHPInsight\Sentiment
+	 */
 	private $sentiment;
+
 	const NEGATIVE = 'negative';
 	const NEUTRAL  = 'neutral';
 	const POSITIVE = 'positive';
 
 	public function __construct()
-	{	
+	{
 		$this->sentiment = new Sentiment();
 	}
 
 	/**
 	 * Get the sentiment of a phrase
+	 *
 	 * @param  string $string The given sentence
 	 * @return string Possible values: negative|neutral|positive
 	 */
@@ -23,24 +28,23 @@ class SentimentAnalysis {
 	{
 		// Do not call functions so that we'll compute only one time
 		$dominantClass = $this->sentiment->categorise($string);
-		
-		switch ($dominantClass) {
+
+		switch ($dominantClass)
+		{
 			case 'neg':
 				return self::NEGATIVE;
-				break;
 
 			case 'neu':
 				return self::NEUTRAL;
-				break;
 
 			case 'pos':
 				return self::POSITIVE;
-				break;
 		}
 	}
 
 	/**
 	 * Get scores for each decision
+	 *
 	 * @param  string $string The original string
 	 * @return array  An array containing keys 'negative', 'neutral' and 'positive' with a float. The closer to 1, the better
 	 * @example ['negative' => 0.5, 'neutral' => 0.25, 'positive' => 0.25]
@@ -54,23 +58,26 @@ class SentimentAnalysis {
 		// We will remap to 'negative' / 'neutral' / 'positive' and round with 2 digits
 		foreach ([self::NEGATIVE, self::NEUTRAL, self::POSITIVE] as $value)
 			$array[$value] = round($scores[substr($value, 0, 3)], 2);
-		
+
 		return $array;
 	}
 
 	/**
 	 * Get the confidence of a decision for a result. The closer to 1, the better
+	 *
 	 * @param  string $string The given sentence
 	 * @return float The confidence of a decision for a result. The close to 1, the better
 	 */
 	public function score($string)
 	{
 		$scores = $this->scores($string);
+
 		return max($scores);
 	}
 
 	/**
 	 * Tells if a sentence is positive
+	 *
 	 * @param  string $string The given sentence
 	 * @return boolean
 	 */
@@ -81,6 +88,7 @@ class SentimentAnalysis {
 
 	/**
 	 * Tells if a sentence is negative
+	 *
 	 * @param  string $string The given sentence
 	 * @return boolean
 	 */
@@ -91,6 +99,7 @@ class SentimentAnalysis {
 
 	/**
 	 * Tells if a sentence is neutral
+	 *
 	 * @param  string $string The given sentence
 	 * @return boolean
 	 */
